+++ console/__init__.py
@@ -1,8 +1,9 @@
 from flask import Flask
 
-__version__ = "1.3.0"
+__version__ = "2.0.0"
+
 
 app = Flask(__name__)
-app.config['USE_MLSD'] = True
 
 import console.views  # noqa
+from console import authentication  # noqa