+++ console/views.py
@@ -1,322 +1,249 @@
-from flask import request, render_template, jsonify
+import ast
+import json
+import logging
+import math
 
-from console import app
-
-import operator
-from ftplib import FTP
 from datetime import datetime
-from console.encrypter import Encrypter
-from console.queue_publisher import QueuePublisher
-import console.settings as settings
+from flask import jsonify
+from flask import redirect
+from flask import render_template
+from flask import request
+from flask import url_for
+import flask_security
+import requests
+from sqlalchemy import func
+from sqlalchemy.exc import DataError, SQLAlchemyError
 from structlog import wrap_logger
 
-import os
-import json
-import requests
-import logging
-import logging.handlers
-import uuid
+from console import app
+from console import settings
+from console.database import db, SurveyResponse, create_dev_user
+from console.forms import NewUserForm
+from console.helpers.exceptions import ClientError, ResponseError, ServiceError, UserCreationError, UserExistsError
 
-from flask_paginate import Pagination
-
-
-PATHS = {
-    "pck": settings.SDX_FTP_DATA_PATH,
-    "image": settings.SDX_FTP_IMAGE_PATH + "/Images",
-    "index": settings.SDX_FTP_IMAGE_PATH + "/Index",
-    "receipt": settings.SDX_FTP_RECEIPT_PATH
-}
 
 logger = wrap_logger(logging.getLogger(__name__))
 
 
-class ConsoleFtp(object):
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        self._ftp.quit()
-
-    def __init__(self):
-        self._ftp = FTP(settings.FTP_HOST)
-        self._ftp.login(user=settings.FTP_USER, passwd=settings.FTP_PASS)
-        self._mlsd_enabled = True
-        try:
-            # Perform a simple mlsd test to see if the ftp server has the extra functionality:
-            len([fname for fname, fmeta in self._ftp.mlsd(path=PATHS['pck'])])
-        except Exception as e:
-            logger.error("Exception initialising consoleftp", exception=e)
-            app.config['USE_MLSD'] = False
-            self._mlsd_enabled = False
-
-    def get_folder_contents(self, path):
-
-        file_list = []
-
-        if self._mlsd_enabled:
-            for fname, fmeta in self._ftp.mlsd(path=path):
-                if fname not in ('.', '..', '.DS_Store'):
-                    meta = {
-                        'name': fname,
-                        'modify': datetime.strptime(fmeta['modify'], '%Y%m%d%H%M%S').isoformat(),
-                        'size': fmeta['size']
-                    }
-                    file_list.append(meta)
-
-        else:
-            pre = []
-            self._ftp.dir("{}".format(path), pre.append)
-            for unparsed_line in pre:
-                bits = unparsed_line.split()
-                date_string = ' '.join([bits[0], bits[1]])
-                fname = ' '.join(bits[3:])
-                # the isdigit() checks this is a file and a directory
-                if fname not in ('.', '..', '.DS_Store') and bits[2].isdigit():
-                    meta = {
-                        'name': fname,
-                        'modify': datetime.strptime(date_string, '%m-%d-%y %I:%M%p').isoformat(),
-                        'size': int(bits[2])
-                    }
-                    file_list.append(meta)
-
-        file_list.sort(key=operator.itemgetter('modify'), reverse=True)
-        return file_list
-
-    def get_file_contents(self, datatype, filename):
-        self._ftp.retrbinary("RETR " + PATHS[datatype] + "/" + filename, open('tmpfile', 'wb').write)
-        transferred = open('tmpfile', 'r')
-        return transferred.read()
+@app.errorhandler(ResponseError)
+def handle_invalid_usage(error):
+    json_error = {"message": error.message, "status_code": error.status_code}
+    return jsonify(json_error)
 
 
-def get_ftp_contents():
-
-    ftp_data = {}
-    with ConsoleFtp() as ftp:
-        ftp_data["pck"] = ftp.get_folder_contents(PATHS["pck"])[0:10]
-        ftp_data["index"] = ftp.get_folder_contents(PATHS["index"])[0:10]
-        ftp_data["image"] = ftp.get_folder_contents(PATHS["image"])[0:10]
-        ftp_data["receipt"] = ftp.get_folder_contents(PATHS["receipt"])[0:10]
-
-    return ftp_data
+@app.route('/healthcheck', methods=['GET'])
+def healthcheck():
+    return jsonify({'status': 'OK'})
 
 
-def list_surveys():
-    return [f for f in os.listdir('console/static/surveys') if os.path.isfile(os.path.join('console/static/surveys', f))]
+@app.route('/', methods=['GET'])
+def home():
+    return render_template('home.html',
+                           current_user=flask_security.core.current_user,
+                           request=request)
 
 
-@app.route('/surveys')
-def surveys():
-    return json.dumps(list_surveys(), indent=4)
+def send_data(logger, url, data=None, json=None, request_type="POST"):
+    response_logger = logger.bind(url=url)
+    try:
+        if request_type == "POST":
+            response_logger.info("Sending POST request")
+            if data:
+                r = requests.post(url, data=data)
+            elif json:
+                r = requests.post(url, json=json)
+        elif request_type == "GET":
+            response_logger.info("Sending GET request")
+            r = requests.get(url)
+    except requests.exceptions.ConnectionError as e:
+        response_logger.error("Failed to connect to service", response="Connection error")
+        raise e
 
+    if 199 < r.status_code < 300:
+        logger.info('Returned from service', response=r.reason, status_code=r.status_code)
+    elif 399 < r.status_code < 500:
+        logger.error('Returned from service', response=r.reason, status_code=r.status_code)
+        raise ClientError(url=url, message=r.reason, status_code=r.status_code)
+    elif r.status_code > 499:
+        logger.error('Returned from service', response=r.reason, status_code=r.status_code)
+        raise ServiceError(url=url, message=r.reason, status_code=r.status_code)
 
-@app.route('/surveys/<survey_id>')
-def survey(survey_id):
-    with open("console/static/surveys/%s" % survey_id) as json_file:
-        file_content = json_file.read()
-        json_file.close()
-        return file_content
-
-
-def send_payload(payload, tx_id, no_of_submissions=1):
-    logger.debug(" [x] Sending encrypted Payload")
-
-    publisher = QueuePublisher(logger, settings.RABBIT_URLS, settings.RABBIT_QUEUE)
-    for i in range(no_of_submissions):
-        publisher.publish_message(payload, tx_id)
-
-    logger.debug(" [x] Sent Payload to rabbitmq!")
-
-
-def mod_to_iso(file_modified):
-    t = datetime.strptime(file_modified, '%Y%m%d%H%M%S')
-    return t.isoformat()
-
-
-def get_image(filename):
-
-    filepath, ext = os.path.splitext(filename)
-
-    tmp_image_url = 'static/images/' + filepath + ext
-    tmp_image_path = 'console/static/images/' + filepath + ext
-
-    if os.path.exists(tmp_image_path):
-        os.unlink(tmp_image_path)
-
-    with ConsoleFtp() as ftp:
-        ftp._ftp.retrbinary("RETR " + PATHS['image'] + "/" + filename, open(tmp_image_path, 'wb').write)
-
-    return tmp_image_url
-
-
-def get_file_contents(datatype, filename):
-
-    with ConsoleFtp() as ftp:
-        return ftp.get_file_contents(datatype, filename)
-
-
-@app.route('/', methods=['POST', 'GET'])
-def submit():
-
-    if request.method == 'POST':
-
-        data = request.get_data().decode('UTF8')
-
-        logger.debug(" [x] Encrypting data")
-
-        unencrypted_json = json.loads(data)
-
-        no_of_submissions = int(unencrypted_json['quantity'])
-
-        encrypter = Encrypter()
-
-        for i in range(0, no_of_submissions):
-            # If submitting more than one then randomise the tx_id
-            if no_of_submissions > 1:
-                tx_id = str(uuid.uuid4())
-                unencrypted_json['survey']['tx_id'] = tx_id
-                logger.info("Auto setting tx_id", tx_id=tx_id)
-            else:
-                tx_id = unencrypted_json['survey']['tx_id']
-
-            payload = encrypter.encrypt(unencrypted_json['survey'])
-            send_payload(payload, tx_id, 1)  # let the loop handle the submission
-
-        return data
-    else:
-
-        return render_template('index.html',
-                               enable_empty_ftp=settings.ENABLE_EMPTY_FTP)
-
-
-def client_error(error=None):
-    logger.error(error)
-    message = {
-        'status': 400,
-        'message': error,
-        'uri': request.url
-    }
-    resp = jsonify(message)
-    resp.status_code = 400
-
-    return resp
-
-
-def get_paginate_info(ru_ref):
-    # This is a little hacky as pagination.start or
-    # pagination.end does not work in the view.
-    info = "Found <b>{total}</b>,"
-    if ru_ref:
-        info += " matching ru_ref: <b>{0}</b>,".format(ru_ref)
-    info += " displaying <b>{start} - {end}</b>"
-    return info
-
-
-@app.route('/store', methods=['POST', 'GET'])
-def store(invalid=False):
-    if request.method == 'POST':
-        mongo_id = request.get_data().decode('UTF8')
-        result = requests.post(settings.SDX_STORE_URL + 'queue', json={"id": mongo_id})
-        return mongo_id if result.status_code is 200 else result
-
-    else:
-        return fetch_responses_from_store(invalid=False)
-
-
-@app.route('/store/invalid', methods=['GET'])
-def store_invalid():
-    return fetch_responses_from_store(invalid=True)
-
-
-def fetch_responses_from_store(invalid=False):
-    params = {}
-    params['page'] = request.args.get('page', type=int, default=1)
-    params['per_page'] = request.args.get('per_page', type=int, default=25)
-    params['ru_ref'] = request.args.get('ru_ref', type=str, default="")
-
-    request_url = '{}{}'.format(settings.SDX_STORE_URL, 'responses')
-    if invalid:
-        invalid = True
-        request_url = '{}{}'.format(settings.SDX_STORE_URL, 'invalid-responses')
-    result = requests.get(request_url, params)
-    content = result.content.decode('UTF8')
-    data = json.loads(content)
-    count = data['total_hits']
-
-    display = get_paginate_info(params['ru_ref'])
-    pagination = Pagination(page=params['page'], total=count, record_name='submissions',
-                            css_framework='bootstrap3', per_page=params['per_page'],
-                            display_msg=display)
-    return render_template('store.html', data=data,
-                           ru_ref=params['ru_ref'], pagination=pagination, invalid=invalid)
+    return r
 
 
 @app.route('/decrypt', methods=['POST', 'GET'])
+@flask_security.login_required
 def decrypt():
-    if request.method == 'POST':
+    audited_logger = logger.bind(user=flask_security.core.current_user.email)
+    if request.method == "POST":
+        data = request.form.get('EncryptedData')
+        url = settings.SDX_DECRYPT_URL
+        decrypted_data = ""
 
-        payload = request.get_data().decode('UTF8')
-
-        send_payload(payload, tx_id=None)
-
-        return payload
-    else:
-
-        ftp_data = get_ftp_contents()
-
-        return render_template('decrypt.html', ftp_data=json.dumps(ftp_data))
-
-
-@app.route('/validate', methods=['POST', 'GET'])
-def validate():
-    if request.method == 'POST':
-
-        payload = request.get_data()
-
-        logger.debug("Validating json...")
-
-        logger.debug("Validate URL: {}".format(settings.SDX_VALIDATE_URL))
-
-        r = requests.post(settings.SDX_VALIDATE_URL, data=payload)
-
-        return jsonify(json.loads(r.text))
-    else:
-
-        ftp_data = get_ftp_contents()
-
-        return render_template('decrypt.html', ftp_data=json.dumps(ftp_data))
-
-
-@app.route('/ftp.json')
-def ftp_list():
-    return jsonify(get_ftp_contents())
-
-
-@app.route('/view/<datatype>/<filename>')
-def view_file(datatype, filename):
-    if filename.lower().endswith(('jpg', 'png')):
-        return '<img style="width:100%;" src="/' + get_image(filename) + '" />'
-    else:
-        return '<pre>' + get_file_contents(datatype, filename) + '</pre>'
-
-
-@app.route('/clear')
-def clear():
-    removed = 0
-
-    with ConsoleFtp() as ftp:
-
-        if app.config['USE_MLSD']:
-            for key, path in PATHS.items():
-                for fname, fmeta in ftp._ftp.mlsd(path=path):
-                    if fname not in ('.', '..'):
-                        ftp._ftp.delete(path + "/" + fname)
-                        removed += 1
+        try:
+            audited_logger.info("Posting data to sdx-decrypt")
+            decrypt_response = send_data(logger=audited_logger, url=url, data=data, request_type="POST")
+        except ClientError:
+            error = 'Client error'
+        except ServiceError:
+            error = 'Service error'
+        except requests.exceptions.ConnectionError:
+            error = 'Connection error'
         else:
-            for key, path in PATHS.items():
-                for fname, fmeta in ftp._ftp.nlst(path):
-                    if fname not in ('.', '..'):
-                        ftp._ftp.delete(path + "/" + fname)
-                        removed += 1
+            decrypted_data = decrypt_response.text
+            error = ""
 
-        return json.dumps({"removed": removed})
+        return render_template('decrypt.html',
+                               decrypted_data=decrypted_data,
+                               error=error,
+                               current_user=flask_security.core.current_user)
+
+    else:
+        return render_template('decrypt.html', current_user=flask_security.core.current_user)
+
+
+def get_filtered_responses(logger, valid, tx_id, ru_ref, survey_id, datetime_earliest, datetime_latest):
+    logger.info('Retrieving responses from sdx-store')
+    try:
+        q = db.session.query(SurveyResponse)
+        if valid == "invalid":
+            q = q.filter(SurveyResponse.invalid)
+        elif valid == "valid":
+            q = q.filter(not SurveyResponse.invalid)
+        if tx_id != '':
+            q = q.filter(SurveyResponse.tx_id == tx_id)
+        if ru_ref != '':
+            q = q.filter(SurveyResponse.data["metadata"]["ru_ref"].astext == ru_ref)
+        if survey_id != '':
+            q = q.filter(SurveyResponse.data["survey_id"].astext == survey_id)
+        dt_column = func.date(SurveyResponse.data["submitted_at"].astext)
+        if datetime_earliest:
+            year = int(datetime_earliest[:4])
+            month = int(datetime_earliest[5:7])
+            day = int(datetime_earliest[8:10])
+            hour = int(datetime_earliest[11:13])
+            minute = int(datetime_earliest[14])
+            q = q.filter(dt_column > datetime(year, month, day, hour, minute))
+        if datetime_latest:
+            year = int(datetime_latest[:4])
+            month = int(datetime_latest[5:7])
+            day = int(datetime_latest[8:10])
+            hour = int(datetime_latest[11:13])
+            minute = int(datetime_latest[14])
+            q = q.filter(dt_column < datetime(year, month, day, hour, minute))
+        filtered_data = q.all()
+    except DataError as e:
+        logger.error("Invalid search term", error=e)
+        return []
+    except SQLAlchemyError as e:
+        logger.error("Database error", error=e)
+        return []
+
+    return filtered_data
+
+
+def reprocess_transaction(logger, json_data):
+    logger.info("Reprocessing transaction", tx_id=json_data["tx_id"])
+    if json_data.get("invalid"):
+        del json_data["invalid"]
+    validate_response = send_data(logger=logger, url=settings.SDX_VALIDATE_URL, json=json_data, request_type="POST")
+    if validate_response != 200:
+        json_data['invalid'] = "True"
+    send_data(logger=logger, url=settings.SDX_STORE_URL, json=json_data, request_type="POST")
+
+
+@app.route('/store/', defaults={'page': 0}, methods=['GET', 'POST'])
+@app.route('/store/<page>', methods=['GET', 'POST'])
+@flask_security.login_required
+def store(page):
+    audited_logger = logger.bind(user=flask_security.core.current_user.email)
+    if request.method == 'POST':
+        json_survey_data = request.form.get('json_data')
+        if not json_survey_data:
+            json_survey_data = request.form.get('json_data_list')
+            if not json_survey_data:
+                return url_for('store')
+            json_survey_data = ast.literal_eval(json_survey_data)
+
+        if isinstance(json_survey_data, list):
+            audited_logger.info("Reprocessing multiple transactions")
+            for json_data in json_survey_data:
+                reprocess_transaction(audited_logger, json_data)
+        else:
+            json_single_data = json.loads(json_survey_data.replace("'", '"'))
+            reprocess_transaction(audited_logger, json_single_data)
+        return redirect(url_for('store'))
+
+    else:
+        valid = request.args.get('valid', type=str, default='')
+        tx_id = request.args.get('tx_id', type=str, default='')
+        ru_ref = request.args.get('ru_ref', type=str, default='')
+        survey_id = request.args.get('survey_id', type=str, default='')
+        datetime_earliest = request.args.get('datetime_earliest', type=str, default='')
+        datetime_latest = request.args.get('datetime_latest', type=str, default='')
+        store_data = get_filtered_responses(audited_logger, valid, tx_id, ru_ref, survey_id, datetime_earliest, datetime_latest)
+        audited_logger.info("Successfully retireved responses")
+        json_list = [item.data for item in store_data]
+        no_pages = math.ceil(round(float(len(json_list) / 20)))
+
+        return render_template('store.html',
+                               data=json_list,
+                               no_pages=no_pages,
+                               page=int(page),
+                               current_user=flask_security.core.current_user)
+
+
+@app.route('/storetest', methods=['GET'])
+def storetest():
+
+    def create_test_data(number):
+        test_data = json.dumps(
+            {
+                "collection": {
+                    "exercise_sid": "hfjdskf",
+                    "instrument_id": "ce2016",
+                    "period": "0616"
+                },
+                "data": {
+                    "1": "2",
+                    "2": "4",
+                    "3": "2",
+                    "4": "Y"
+                },
+                "metadata": {
+                    "ru_ref": "12345678901a",
+                    "user_id": "789473423"
+                },
+                "origin": "uk.gov.ons.edc.eq",
+                "submitted_at": "2017-04-27T14:23:13+00:00",
+                "survey_id": "1",
+                "tx_id": "f088d89d-a367-876e-f29f-ae8f1a26" + str(number),
+                "type": "uk.gov.ons.edc.eq:surveyresponse",
+                "version": "0.0.1"
+            })
+        return test_data
+    for i in range(1000, 1250):
+        test_data = create_test_data(str(i))
+        send_data(logger=logger, url=settings.SDX_STORE_URL, data=test_data, request_type="POST")
+
+    return "data sent"
+
+
+@app.route('/adduser', methods=['GET', 'POST'])
+@flask_security.roles_required('Admin')
+def add_user():
+    form = NewUserForm()
+    if request.method == 'POST' and form.validate():
+        success = False
+        try:
+            create_dev_user(form.email.data, form.password.data)
+            success = True
+        except UserCreationError:
+            form.errors['Database'] = ["Error creating user"]
+        except UserExistsError:
+            form.errors['User'] = ["This user already exists"]
+
+        return render_template('adduser.html', form=form, success=success, user=form.email.data)
+
+    else:
+        return render_template('adduser.html', form=form)